move method to s3.py + typo + default value
